drivers: wakeup: add entries to disable timerfd, netlink, wlan_wow_wl, wlan and netmgr_wl wakelocks
As usual it's the user's entire responsibility of breaking (or not) functionality when disabling wakelocks.
wlan_wow_wl, wlan and netmgr_wl wakelocks are disabled by default. The other two are not since it's the first
time I'm allowing them to be disabled on any device and I don't know what (if anything) can happen.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>